Refactored tests according to the agreement
