+++ src/test/java/com/coherentsolutions/pot/insurance_service/integration/InsuranceServiceApplicationTests.java
@@ -1,9 +1,10 @@
-package com.coherentsolutions.pot.insurance_service;
+package com.coherentsolutions.pot.insurance_service.integration;
 
-import com.coherentsolutions.pot.insurance_service.containers.PostgresTestContainer;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestPropertySource;
+
+import com.coherentsolutions.pot.insurance_service.integration.containers.PostgresTestContainer;
 
 @SpringBootTest
 @TestPropertySource(properties = "spring.main.allow-bean-definition-overriding=true")
